requirements.txt
pandas
scikit-learn
torchvision
numpy
matplotlib